move bus connect method to process_utils
